+++ leda/components/NumericRange/index.tsx
@@ -29,6 +29,7 @@ export const NumericRange = React.forwardRef((props: NumericRangeProps, ref?: Re
     name: nameProp,
     onBlur,
     onChange,
+    onEnterPress,
     onFocus,
     placeholder: placeholderProp,
     shouldTrimTrailingZeros,
@@ -106,6 +107,7 @@ export const NumericRange = React.forwardRef((props: NumericRangeProps, ref?: Re
         max={isNil(value[1]) ? max : value[1]}
         min={min}
         name={name[0]}
+        onEnterPress={onEnterPress}
         onChange={handleChange('from')}
         placeholder={placeholder[0]}
         theme={theme.to}
@@ -120,6 +122,7 @@ export const NumericRange = React.forwardRef((props: NumericRangeProps, ref?: Re
         max={max}
         min={isNil(value[0]) ? min : value[0]}
         name={name[1]}
+        onEnterPress={onEnterPress}
         onChange={handleChange('to')}
         placeholder={placeholder[1]}
         theme={theme.from}